Refactor: test/e2e 폴더 추가 및 구조 변경
